requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pyotp
python-jose
email-validator
passlib
cloudinary
